Conflicts:
	src/main/java/com/brewduck/web/common/dao/FileDaoImpl.java

+++ src/main/java/com/brewduck/web/common/dao/FileDaoImpl.java
@@ -37,20 +37,17 @@ public class FileDaoImpl implements FileDao {
 
     @Override
     public  int selectFileSeq() {
-       return sqlSessionTemplate.selectOne("Common.selectFileSeq");
-       // return 1;
+        return (Integer)sqlSessionTemplate.selectOne("Common.selectFileSeq");
     }
 
     @Override
     public  int selectFileNo(Integer seq) {
-       return sqlSessionTemplate.selectOne("Common.selectFileNo",seq);
-       // return 1;
+        return (Integer)sqlSessionTemplate.selectOne("Common.selectFileNo",seq);
     }
 
     @Override
     public  int selectFileCount(Integer seq) {
-        return sqlSessionTemplate.selectOne("Common.selectFileCount", seq);
-       // return 1;
+        return (Integer)sqlSessionTemplate.selectOne("Common.selectFileCount", seq);
     }
 
     @Override
